BugFix > null string input Value Error at Integer field
The cause of the error is that Django tries to insert null-string value at Integer type field in DB.

According to StackOverflow,
If the field is passed in None, It will be translated to Null.

However, my Django seems to read null-string('') but 'None' Why?

```
form.aged = request.POST.get('input_aged') if request.POST.get('input_aged') else None
```

I fixed the error as this code eventually, but I am not satisfied with this way.